+++ ChatServices/MatrixChatServiceConnection.cs
@@ -13,6 +13,7 @@ using System.Web;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Theorem.Models;
+using Theorem.Models.Matrix;
 
 namespace Theorem.ChatServices
 {
@@ -302,7 +303,8 @@ namespace Theorem.ChatServices
             onNewMessage(messageModel);
         }
 
-        private async Task<MatrixSendResponse> sendTextMessageToRoomAsync(string roomId, string body)
+        private async Task<MatrixSendResponse> sendTextMessageToRoomAsync(string roomId,
+            string body)
         {
             // https://matrix.org/docs/spec/client_server/latest#id271
             var payload = new Dictionary<string, string>()
@@ -342,276 +344,6 @@ namespace Theorem.ChatServices
             if (eventHandler != null)
             {
                 eventHandler(this, message);
-            }
-        }
-
-        private record MatrixError
-        {
-            [JsonPropertyName("errcode")]
-            public string ErrorCode { get; init; }
-
-            [JsonPropertyName("error")]
-            public string Error { get; init; }
-        }
-
-        private record MatrixLoginIdentifier
-        {
-            [JsonPropertyName("type")]
-            public string Type { get; init; }
-
-            [JsonPropertyName("user")]
-            public string User { get; init; }
-        }
-
-        private record MatrixLoginPayload
-        {
-            [JsonPropertyName("type")]
-            public string Type { get; init; }
-
-            [JsonPropertyName("identifier")]
-            public MatrixLoginIdentifier Identifier { get; init; }
-
-            [JsonPropertyName("password")]
-            public string Password { get; init; }
-
-            [JsonPropertyName("device_id")]
-            public string DeviceId { get; init; }
-
-            [JsonPropertyName("initial_device_display_name")]
-            public string InitialDeviceDisplayName { get; init; }
-        }
-
-        private record MatrixLoginResponse
-        {
-            [JsonPropertyName("user_id")]
-            public string UserId { get; init; }
-
-            [JsonPropertyName("access_token")]
-            public string AccessToken { get; init; }
-
-            [JsonPropertyName("device_id")]
-            public string DeviceId { get; init; }
-
-            // additional fields @ https://matrix.org/docs/spec/client_server/latest#id205
-        }
-
-        private record MatrixJoinRoomResponse
-        {
-            [JsonPropertyName("room_id")]
-            public string RoomId { get; init; }
-        }
-
-        private record MatrixSendResponse
-        {
-            [JsonPropertyName("event_id")]
-            public string EventId { get; init; }
-        }
-
-        private record MatrixDirectoryRoomResponse
-        {
-            [JsonPropertyName("room_id")]
-            public string RoomId { get; init; }
-
-            // Additional fields @ https://matrix.org/docs/spec/client_server/latest#id282
-        }
-
-        private record MatrixSyncResponse
-        {
-            [JsonPropertyName("next_batch")]
-            public string NextBatchToken { get; init; }
-
-            [JsonPropertyName("rooms")]
-            public MatrixSyncResponseRooms Rooms { get; init; }
-
-            // Additional fields @ https://matrix.org/docs/spec/client_server/latest#id257
-        }
-
-        private record MatrixSyncResponseRooms
-        {
-            [JsonPropertyName("join")]
-            public Dictionary<string, MatrixJoinedRoom> JoinedRooms { get; init; }
-
-            [JsonPropertyName("invite")]
-            public Dictionary<string, MatrixInvitedRoom> InvitedRooms { get; init; }
-
-            [JsonPropertyName("leave")]
-            public Dictionary<string, MatrixLeftRoom> LeftRooms { get; init; }
-        }
-
-        private record MatrixJoinedRoom
-        {
-            [JsonPropertyName("state")]
-            public MatrixState State { get; init; }
-
-            [JsonPropertyName("timeline")]
-            public MatrixTimeline Timeline { get; init; }
-
-            // Additional fields https://matrix.org/docs/spec/client_server/latest#id257
-        }
-
-        private record MatrixState
-        {
-            [JsonPropertyName("events")]
-            public IList<MatrixEvent> Events { get; init; }
-        }
-
-        private record MatrixTimeline
-        {
-            [JsonPropertyName("events")]
-            public IList<MatrixEvent> Events { get; init; }
-
-            [JsonPropertyName("limited")]
-            public bool IsLimited { get; init; }
-
-            [JsonPropertyName("prev_batch")]
-            public string PreviousBatchToken { get; init; }
-        }
-
-        private record MatrixInvitedRoom
-        {
-            [JsonPropertyName("invite_state")]
-            public MatrixInviteState InviteState { get; init; }
-        }
-
-        private record MatrixInviteState
-        {
-            [JsonPropertyName("events")]
-            public IList<MatrixStrippedState> Events { get; init; }
-        }
-
-        private record MatrixLeftRoom
-        {
-            [JsonPropertyName("state")]
-            public MatrixState State { get; init; }
-
-            [JsonPropertyName("timeline")]
-            public MatrixTimeline Timeline { get; init; }
-
-            // Additional fields
-        }
-
-        private record MatrixStrippedState
-        {
-            [JsonPropertyName("content")]
-            public dynamic EventContent { get; init; }
-
-            [JsonPropertyName("state_key")]
-            public string StateKey { get; init; }
-
-            [JsonPropertyName("type")]
-            public string Type { get; init; }
-
-            [JsonPropertyName("sender")]
-            public string Sender { get; init; }
-        }
-
-        [JsonConverter(typeof(MatrixEventJsonConverter))]
-        private record MatrixEvent
-        {
-            // https://matrix.org/docs/spec/client_server/latest#id244
-
-            [JsonPropertyName("type")]
-            public string Type { get; init; }
-        }
-
-        private record MatrixRoomEvent : MatrixEvent
-        {
-            // m.room
-            // https://matrix.org/docs/spec/client_server/latest#id245
-
-            [JsonPropertyName("event_id")]
-            public string EventId { get; init; }
-
-            [JsonPropertyName("sender")]
-            public string Sender { get; init; }
-
-            [JsonPropertyName("origin_server_ts")]
-            public UInt64 OriginServerTimestamp { get; init; }
-
-            // TODO: Unsigned data
-
-            [JsonPropertyName("room_id")]
-            public string RoomId { get; init; }
-        }
-
-        private record MatrixRoomMessageEvent : MatrixRoomEvent
-        {
-            [JsonPropertyName("content")]
-            public MatrixRoomMessageEventContent Content { get; init; }
-        }
-
-        private record MatrixRoomMessageEventContent
-        {
-            [JsonPropertyName("body")]
-            public string Body { get; init; }
-
-            [JsonPropertyName("msgtype")]
-            public string MessageType { get; init; }
-        }
-
-        private record MatrixRoomMemberEvent : MatrixRoomEvent
-        {
-            [JsonPropertyName("content")]
-            public MatrixRoomMemberEventContent Content { get; init; }
-            
-            [JsonPropertyName("state_key")]
-            public string StateKey { get; init; }
-        }
-
-        private record MatrixRoomMemberEventContent
-        {
-            // https://matrix.org/docs/spec/client_server/latest#id252
-
-            [JsonPropertyName("avatar_url")]
-            public string AvatarUrl { get; init; }
-
-            [JsonPropertyName("displayname")]
-            public string DisplayName { get; init; }
-
-            [JsonPropertyName("membership")]
-            public string Membership { get; init; }
-
-            [JsonPropertyName("is_direct")]
-            public bool IsDirect { get; init; }
-
-            // Additional fields 
-        }
-
-        private class MatrixEventJsonConverter : JsonConverter<MatrixEvent>
-        {
-            public override bool CanConvert(Type typeToConvert)
-            {
-                return typeToConvert.IsAssignableFrom(typeof(MatrixEvent));
-            }
-
-            public override MatrixEvent Read(ref Utf8JsonReader reader, Type typeToConvert,
-                JsonSerializerOptions options)
-            {
-                if (JsonDocument.TryParseValue(ref reader, out var doc))
-                {
-                    if (doc.RootElement.TryGetProperty("type", out var type))
-                    {
-                        var typeValue = type.GetString();
-                        var rootElement = doc.RootElement.GetRawText();
-
-                        return typeValue switch
-                        {
-                            "m.room.message" => JsonSerializer
-                                .Deserialize<MatrixRoomMessageEvent>(rootElement, options),
-                            "m.room.member" => JsonSerializer
-                                .Deserialize<MatrixRoomMemberEvent>(rootElement, options),
-                            _ => new MatrixEvent() { Type = typeValue },
-                        };
-                    }
-                    throw new JsonException("Failed to extract type property from MatrixEvent");
-                }
-                throw new JsonException("Failed to parse MatrixEvent");
-            }
-
-            public override void Write(Utf8JsonWriter writer, MatrixEvent value,
-                JsonSerializerOptions options)
-            {
-                throw new NotImplementedException();
             }
         }
     }
